Fix Klarna integration (CS1)
Existing implementation didn't include environment specific base URL applicable
to Klarna client operations. This commit adds the ability to use the `SANDBOX`
base url and integration tests for create and update session scenarios.

It also fixes mapping issues in request/response classes. In fields related to
amounts, Integers were replaced by Long types.

+++ src/main/java/com/checkout/apm/klarna/KlarnaProduct.java
@@ -10,18 +10,18 @@ public final class KlarnaProduct {
 
     private String name;
 
-    private Integer quantity;
+    private Long quantity;
 
     @SerializedName("unit_price")
-    private Integer unitPrice;
+    private Long unitPrice;
 
     @SerializedName("tax_rate")
-    private Integer taxRate;
+    private Long taxRate;
 
     @SerializedName("total_amount")
-    private Integer totalAmount;
+    private Long totalAmount;
 
     @SerializedName("total_tax_amount")
-    private Integer totalTaxAmount;
+    private Long totalTaxAmount;
 
 }
